+++ osconfiglib/cli/main.py
@@ -67,5 +67,13 @@ def delete_layer(layer_name):
     click.echo(f'Deleting layer {layer_name}.')
 cli.add_command(delete_layer, name='delete')
 
+
+@click.command()
+@click.argument('recipe')
+def export_squashed_configs(recipe):
+    # Here you would call the functionality that deletes a layer
+    click.echo(f'Squashing configs for {recipe}.')
+cli.add_command(export_squashed_configs, name='export-squashed-configs')
+
 if __name__ == '__main__':
     cli()